pyproject.toml
[tool.poetry]
name = "1lb"
version = "0.1.0"
description = ""
authors = ["Maks1mka <maksim_goncharov_2006@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pyserial = "==3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
